+++ docs/.readthedocs.yaml
@@ -20,4 +20,7 @@ sphinx:
 # Optionally declare the Python requirements required to build your docs
 python:
   install:
-    - requirements: docs/.sphinx/requirements.txt
+    - requirements: docs/requirements.txt
+    - method: pip
+      path: .
+
