--- docs/.sphinx/_static/furo_colors.css
@@ -1,88 +1,0 @@
-body {
-    --color-code-background: #f8f8f8;
-    --color-code-foreground: black;
-    --font-stack: Ubuntu, -apple-system, Segoe UI, Roboto, Oxygen, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
-    --font-stack--monospace: Ubuntu Mono, Consolas, Monaco, Courier, monospace;
-    --color-foreground-primary: #111;
-    --color-foreground-secondary: var(--color-foreground-primary);
-    --color-foreground-muted: #333;
-    --color-background-secondary: #FFF;
-    --color-background-hover: #f2f2f2;
-    --color-brand-primary: #111;
-    --color-brand-content: #06C;
-    --color-api-background: #cdcdcd;
-    --color-inline-code-background: rgba(0,0,0,.03);
-    --color-sidebar-link-text: #111;
-    --color-sidebar-item-background--current: #ebebeb;
-    --color-sidebar-item-background--hover: #f2f2f2;
-    --toc-font-size: var(--font-size--small);
-    --color-admonition-title-background--note: var(--color-background-primary);
-    --color-admonition-title-background--tip: var(--color-background-primary);
-    --color-admonition-title-background--important: var(--color-background-primary);
-    --color-admonition-title-background--caution: var(--color-background-primary);
-    --color-admonition-title--note: #24598F;
-    --color-admonition-title--tip: #24598F;
-    --color-admonition-title--important: #C7162B;
-    --color-admonition-title--caution: #F99B11;
-    --color-highlighted-background: #EbEbEb;
-    --color-link-underline: var(--color-background-primary);
-    --color-link-underline--hover: var(--color-background-primary);
-    --color-version-popup: #772953;
-}
-
-@media not print {
-    body[data-theme="dark"] {
-        --color-code-background: #202020;
-        --color-code-foreground: #d0d0d0;
-        --color-foreground-secondary: var(--color-foreground-primary);
-        --color-foreground-muted: #CDCDCD;
-        --color-background-secondary: var(--color-background-primary);
-        --color-background-hover: #666;
-        --color-brand-primary: #fff;
-        --color-brand-content: #06C;
-        --color-sidebar-link-text: #f7f7f7;
-        --color-sidebar-item-background--current: #666;
-        --color-sidebar-item-background--hover: #333;
-        --color-admonition-background: transparent;
-        --color-admonition-title-background--note: var(--color-background-primary);
-        --color-admonition-title-background--tip: var(--color-background-primary);
-        --color-admonition-title-background--important: var(--color-background-primary);
-        --color-admonition-title-background--caution: var(--color-background-primary);
-        --color-admonition-title--note: #24598F;
-        --color-admonition-title--tip: #24598F;
-        --color-admonition-title--important: #C7162B;
-        --color-admonition-title--caution: #F99B11;
-        --color-highlighted-background: #666;
-        --color-link-underline: var(--color-background-primary);
-        --color-link-underline--hover: var(--color-background-primary);
-        --color-version-popup: #F29879;
-    }
-    @media (prefers-color-scheme: dark) {
-        body:not([data-theme="light"]) {
-            --color-code-background: #202020;
-            --color-code-foreground: #d0d0d0;
-            --color-foreground-secondary: var(--color-foreground-primary);
-            --color-foreground-muted: #CDCDCD;
-            --color-background-secondary: var(--color-background-primary);
-            --color-background-hover: #666;
-            --color-brand-primary: #fff;
-            --color-brand-content: #06C;
-            --color-sidebar-link-text: #f7f7f7;
-            --color-sidebar-item-background--current: #666;
-            --color-sidebar-item-background--hover: #333;
-            --color-admonition-background: transparent;
-            --color-admonition-title-background--note: var(--color-background-primary);
-            --color-admonition-title-background--tip: var(--color-background-primary);
-            --color-admonition-title-background--important: var(--color-background-primary);
-            --color-admonition-title-background--caution: var(--color-background-primary);
-            --color-admonition-title--note: #24598F;
-            --color-admonition-title--tip: #24598F;
-            --color-admonition-title--important: #C7162B;
-            --color-admonition-title--caution: #F99B11;
-            --color-highlighted-background: #666;
-            --color-link-underline: var(--color-background-primary);
-            --color-link-underline--hover: var(--color-background-primary);
-            --color-version-popup: #F29879;
-        }
-    }
-}
