--- docs/.sphinx/_templates/base.html
@@ -1,12 +1,0 @@
-{% extends "furo/base.html" %}
-
-{% block theme_scripts %}
-<script>
-  const github_url = "{{ github_url }}";
-</script>
-{% endblock theme_scripts %}
-
-{# ru-fu: don't include the color variables from the conf.py file, but use a
-￼   separate CSS file to save space #}
-{% block theme_styles %}
-{% endblock theme_styles %}
