--- docs/.sphinx/_templates/page.html
@@ -1,49 +1,0 @@
-{% extends "furo/page.html" %}
-
-{% block footer %}
-   {% include "footer.html" %}
-{% endblock footer %}
-
-{% block body -%}
-   {% include "header.html" %}
-   {{ super() }}
-{%- endblock body %}
-
-{% if meta and ((meta.discourse and discourse_prefix) or meta.relatedlinks) %}
-   {% set furo_hide_toc_orig = furo_hide_toc %}
-   {% set furo_hide_toc=false %}
-{% endif %}
-
-{% block right_sidebar %}
-<div class="toc-sticky toc-scroll">
-   {% if not furo_hide_toc_orig %}
-    <div class="toc-title-container">
-      <span class="toc-title">
-       {{ _("Contents") }}
-      </span>
-    </div>
-    <div class="toc-tree-container">
-      <div class="toc-tree">
-        {{ toc }}
-      </div>
-    </div>
-   {% endif %}
-    {% if meta and ((meta.discourse and discourse_prefix) or meta.relatedlinks) %}
-    <div class="relatedlinks-title-container">
-      <span class="relatedlinks-title">
-       Related links
-      </span>
-    </div>
-    <div class="relatedlinks-container">
-      <div class="relatedlinks">
-        {% if meta.discourse and discourse_prefix %}
-          {{ discourse_links(meta.discourse) }}
-        {% endif %}
-        {% if meta.relatedlinks %}
-          {{ related_links(meta.relatedlinks) }}
-        {% endif %}
-      </div>
-    </div>
-    {% endif %}
-  </div>
-{% endblock right_sidebar %}
