--- docs/_static/css/custom.css
@@ -1,28 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Ubuntu:ital@0;1&display=swap');
-
-body {
-    font-family: Ubuntu, "times new roman", times, roman, serif;
-}
-
-div .toctree-wrapper {
-    column-count: 2;
-}
-
-div .toctree-wrapper>ul {
-    margin: 0;
-}
-
-ul .toctree-l1 {
-    margin: 0;
-    -webkit-column-break-inside: avoid;
-    page-break-inside: avoid;
-    break-inside: avoid-column;
-}
-
-.wy-nav-content {
-    max-width: none;
-}
-
-.log-snippets {
-    color: rgb(141, 141, 141);
-}
