+++ tox.ini
@@ -106,7 +106,7 @@ commands =
 
 [docs]  # Sphinx documentation configuration
 deps =
-    -r{tox_root}/docs/.sphinx/requirements.txt
+    -r{tox_root}/docs/requirements.txt
 package = editable
 no_package = true
 env_dir = {work_dir}/docs
